evm/main/requirements.txt
1inch.py==1.9
requests==2.28.1
web3>=6.0.0
boto3==1.26.94
parsimonious==0.9.0
wget==3.2
python-decouple
